cloud/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.19.0
google-cloud-bigquery>=3.27.0
google-cloud-aiplatform>=1.60.0
vertexai>=1.60.0
google-generativeai>=0.8.0
httpx>=0.27.0
